# IntPay.Api/docs/openapi-v1.yaml
openapi: 3.0.3
info:
  title: IntPay API
  version: 1.0.0
  description: |
    OpenAPI documentation for IntPay API v1.

    The current implementation uses explicit caller identifiers (`userId`, `profileId`, or `actingUserId`) as a temporary request-level authorization boundary. JWT/session authentication is not enforced by the API yet.

    In the current MVP, authorized card GET responses may include the full card PAN (`cardNumber`) and CVV (`cvv`) for demo flows. Production deployments should re-enable masking, secret handling, and real authentication.

    `useTimes` is optional when creating an intent. If the client omits it or sends `null`, the API treats the intent as unlimited. The current numeric sentinel for unlimited usage is `99999`; API consumers should display this as `unlimited`.

    In Development, `MapOpenApi()` still exposes generated JSON (e.g. `/openapi/v1.json`). Scalar is configured to load this hand-maintained YAML from `/openapi/intpay-v1.yaml` (same contents as `docs/openapi-v1.yaml` on disk) so examples and long descriptions appear in the UI. Keep this file aligned with `Endpoints/V1` when behavior changes.
servers:
  - url: /api/v1
    description: API v1 relative server
...
paths:
...
  /users/{userId}/activities/latest:
    get:
      tags: [Users]
      summary: Get merged user activity feed
      description: |
        Merges card-scoped `audit_logs` (authorizations, manual freeze, invoice verification, intent updates) for every virtual card where the path user is creator or receiver,
        with profile-scoped rows where `audit_logs.user_id` matches the path user (vault `balance_added`, `balance_deducted`, `wallet_credit`, and intent governance rows on the profile feed).

        Each item exposes derived `entityType` and `outcome` for UI grouping. Use query `status` to filter by normalized outcome; use `decision` for the raw audit decision enum.
      operationId: getLatestUserActivities
      parameters:
        - $ref: '#/components/parameters/UserIdPath'
        - $ref: '#/components/parameters/LimitOptional'
        - $ref: '#/components/parameters/OffsetOptional'
        - name: decision
          in: query
          required: false
          description: Filter by raw audit decision (`approved`, `declined`, `info`).
          schema:
            $ref: '#/components/schemas/DecisionStatus'
          example: approved
        - name: action
          in: query
          required: false
          description: Filter by audit action (e.g. `authorization`, `balance_added`, `card_manual_freeze_set`, `intent_updated`).
          schema:
            type: string
          example: authorization
        - name: entityType
          in: query
          required: false
          description: Filter by derived bucket applied in the API after fetch.
          schema:
            $ref: '#/components/schemas/ActivityEntityType'
          example: transaction
        - name: status
          in: query
          required: false
          description: Filter by normalized outcome; echoed as `filters.outcome` in the response (query name is `status`, not `outcome`).
          schema:
            $ref: '#/components/schemas/ActivityOutcome'
          example: success
        - name: cardId
          in: query
          required: false
          description: Filter by card id.
          schema:
            type: integer
            format: int32
            minimum: 1
          example: 31
        - name: intentId
          in: query
          required: false
          description: Filter by intent id.
          schema:
            type: integer
            format: int32
            minimum: 1
          example: 21
        - name: from
          in: query
          required: false
          description: Inclusive UTC lower bound for created time.
          schema:
            type: string
            format: date-time
          example: '2026-05-01T00:00:00Z'
        - name: to
          in: query
          required: false
          description: Inclusive UTC upper bound for created time.
          schema:
            type: string
            format: date-time
          example: '2026-05-02T00:00:00Z'
        - name: merchant
          in: query
          required: false
          description: Case-insensitive contains on merchant name; for profile-scoped rows, matches `note` or `reason`.
          schema:
            type: string
          example: Taxi
        - name: mcc
          in: query
          required: false
          description: Exact MCC filter.
          schema:
            type: string
          example: '4121'
        - name: city
          in: query
          required: false
          description: Case-insensitive contains filter for city.
          schema:
            type: string
          example: Riyadh
        - name: minAmount
          in: query
          required: false
          description: Minimum transaction amount.
          schema:
            type: number
            format: decimal
            minimum: 0
          example: 10.0
        - name: maxAmount
          in: query
          required: false
          description: Maximum transaction amount.
          schema:
            type: number
            format: decimal
            minimum: 0
          example: 100.0
        - name: role
          in: query
          required: false
          description: User role filter. Unknown values are normalized to `all`.
          schema:
            $ref: '#/components/schemas/ActivityRole'
          example: sender
        - name: includeInfo
          in: query
          required: false
          description: Whether to include informational audit rows.
          schema:
            type: boolean
            default: true
          example: true
      responses:
        '200':
          description: Latest user activities fetched successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserLatestActivitiesEnvelope'
              examples:
                success:
                  value:
                    success: true
                    message: Latest user activities fetched successfully.
                    data:
                      userId: 1
                      total: 1
                      limit: 20
                      offset: 0
                      filters:
                        decision: approved
                        action: null
                        entityType: null
                        outcome: null
                        cardId: null
                        intentId: null
                        fromUtc: null
                        toUtc: null
                        merchant: Taxi
                        mcc: null
                        city: null
                        minAmount: null
                        maxAmount: null
                        role: sender
                        includeInfo: true
                      summary:
                        approvedCount: 1
                        declinedCount: 0
                        infoCount: 0
                        approvedSpendTotal: 25.5
                        declinedAmountTotal: 0.0
                        distinctCards: 1
                        distinctIntents: 1
                      items:
                        - id: 90
                          cardId: 31
                          intentId: 21
                          action: authorization
                          decision: approved
                          reason: null
                          transactionAmount: 25.5
                          merchantName: Riyadh Taxi
                          mcc: '4121'
                          city: Riyadh
                          createdAt: '2026-05-01T02:00:00Z'
                          occurredAt: '2026-05-01T02:00:00Z'
                          creatorId: 1
                          receiverId: 2
                          role: sender
                          intentDescription: Team travel allowance
                          category: null
                          country: SA
                          intentAmount: 750.0
                          remainingAmount: 724.5
                          cardLast4: '1234'
                          cardStatus: active
                          isLockedByPendingInvoice: false
                          isManuallyFrozen: false
                          isSpendBlocked: false
                          senderName: null
                          activityType: authorization
                          title: Approved at Riyadh Taxi
                          subtitle: Team travel allowance · Riyadh · MCC 4121
                          severity: success
                          amountLabel: $25.50
                          entityType: transaction
                          outcome: success
                    meta:
                      statusCode: 200
                      version: v1
                      timestamp: '2026-05-01T02:00:00.0000000+00:00'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /users/{userId}/transactions:
    get:
      tags: [Users]
      summary: Get user transactions
      description: Returns paged audit logs for all cards where the user is creator or receiver.
      operationId: getUserTransactions
      parameters:
        - $ref: '#/components/parameters/UserIdPath'
        - $ref: '#/components/parameters/LimitOptional'
        - $ref: '#/components/parameters/OffsetOptional'
      responses:
        '200':
          description: User transactions fetched successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserTransactionsEnvelope'
              examples:
                success:
                  value:
                    success: true
                    message: User transactions fetched successfully.
                    data:
                      userId: 1
                      total: 1
                      limit: 50
                      offset: 0
                      logs:
                        - id: 90
                          cardId: 31
                          transactionAmount: 25.5
                          merchantName: Riyadh Taxi
                          mcc: '4121'
                          decision: approved
                          reason: null
                          createdAt: '2026-05-01T02:00:00Z'
                          city: Riyadh
                          entityId: 21
                          action: authorization
                          occurredAt: '2026-05-01T02:00:00Z'
                    meta:
                      statusCode: 200
                      version: v1
                      timestamp: '2026-05-01T02:00:00.0000000+00:00'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /users/{userId}/cards/logs-as-receiver:
    get:
      tags: [Users]
      summary: Get receiver card logs
      description: Returns paged audit logs for cards where the user is the intent receiver.
      operationId: getAuditLogsForReceiverCards
      parameters:
        - $ref: '#/components/parameters/UserIdPath'
        - $ref: '#/components/parameters/LimitOptional'
        - $ref: '#/components/parameters/OffsetOptional'
      responses:
        '200':
          description: Receiver card audit logs fetched successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReceiverPagedAuditLogsEnvelope'
              examples:
                success:
                  value:
                    success: true
                    message: Receiver card audit logs fetched successfully.
                    data:
                      receiverUserId: 2
                      total: 1
                      limit: 20
                      offset: 0
                      logs:
                        - id: 90
                          cardId: 31
                          transactionAmount: 25.5
                          merchantName: Riyadh Taxi
                          mcc: '4121'
                          decision: approved
                          createdAt: '2026-05-01T02:00:00Z'
                          city: Riyadh
                    meta:
                      statusCode: 200
                      version: v1
                      timestamp: '2026-05-01T02:00:00.0000000+00:00'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  parameters:
    UserIdPath:
      name: userId
      in: path
      required: true
      description: User profile id.
      schema:
        type: integer
        format: int32
        minimum: 1
      example: 1
...
    LimitOptional:
      name: limit
      in: query
      required: false
      description: Page size. Service clamps supported typed endpoints to 1..1000. Some card endpoints fallback to 50 when empty or invalid.
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 1000
        default: 50
      example: 20
    OffsetOptional:
      name: offset
      in: query
      required: false
      description: Number of records to skip. Negative values are normalized to 0 on typed endpoints.
      schema:
        type: integer
        format: int32
        minimum: 0
        default: 0
      example: 0
  responses:
    BadRequest:
      description: Invalid request body, invalid query/path parameter, validation failure, or business rule error.
      content:
        application/json:
          schema:
            oneOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - $ref: '#/components/schemas/ProblemDetails'
          examples:
            error:
              value:
                success: false
                message: Invalid query parameter value.
                detail: Failed to bind parameter "Nullable<int> limit" from "abc".
            problem:
              value:
                type: about:blank
                title: Bad Request
                status: 400
                detail: Request body is required.
    Unauthorized:
      description: Authentication is not implemented yet, but this standard response is reserved for future JWT/session authentication.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            unauthorized:
              value:
                success: false
                message: Authentication is required.
...
    NotFound:
      description: Requested profile, intent, card, or related resource was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            notFound:
              value:
                success: false
                message: Resource not found.
...
    InternalServerError:
      description: Unexpected server error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            error:
              value:
                success: false
                message: Unexpected server error.
  schemas:
    DecisionStatus:
      type: string
      description: |
        Audit decision status.
        - `approved`: authorization or verification was approved.
        - `declined`: authorization or verification was declined.
        - `info`: non-financial informational event.
      enum: [approved, declined, info]
      example: approved
    ActivityRole:
      type: string
      description: |
        Activity ownership filter (normalized server-side before querying).
        - `all`: all visible activities (default for unknown values).
        - `sender`: creator-only perspective (`creator` and `sent` are treated as `sender`).
        - `receiver`: receiver-only perspective (`recipient` and `received` are treated as `receiver`).
        - `self`: creator and receiver are the same user on the card.
      enum: [all, sender, creator, sent, receiver, recipient, received, self]
      default: all
      example: sender
...
    ActivityEntityType:
      type: string
      description: |
        Derived classification for an activity row (API-computed).
        - `profile`: vault balance changes and similar rows tied to `audit_logs.user_id` without a card.
        - `intent`: intent-level governance events surfaced on the profile feed.
        - `virtual_card`: card-scoped non-authorization events (freeze, invoice verification, intent updates).
        - `transaction`: card authorization attempts (`action = authorization`).
      enum: [profile, intent, virtual_card, transaction]
      example: transaction
    ActivityOutcome:
      type: string
      description: |
        Normalized outcome for filtering and UI (API-computed). Query parameter name is `status`; response echoes `filters.outcome`.
        - `success`: approved authorizations and successful balance credits.
        - `failed`: declined authorizations and failed balance operations when modeled as such.
        - `info`: informational audit rows (non-spend outcomes).
      enum: [success, failed, info]
      example: success
...
    ErrorResponse:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
          description: Always false for this error shape.
          example: false
        message:
          type: string
          description: Human-readable error message.
          example: Query parameter actingUserId is required.
        detail:
          type: string
          nullable: true
          description: Optional technical detail, especially for binding failures.
          example: Failed to bind parameter "Nullable<int> limit" from "abc".
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          example: about:blank
        title:
          type: string
          example: Bad Request
        status:
          type: integer
          example: 400
        detail:
          type: string
          example: Request body is required.
    Meta:
      type: object
      properties:
        statusCode:
          type: integer
          example: 200
        version:
          type: string
          example: v1
        timestamp:
          type: string
          format: date-time
          example: '2026-05-01T02:00:00.0000000+00:00'
...
    UserLatestActivitiesEnvelope:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelopeBase'
        - type: object
          required: [data, meta]
          properties:
            data:
              $ref: '#/components/schemas/UserLatestActivitiesResponse'
            meta:
              $ref: '#/components/schemas/Meta'
    UserTransactionsEnvelope:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelopeBase'
        - type: object
          required: [data, meta]
          properties:
            data:
              $ref: '#/components/schemas/UserTransactionsResponse'
            meta:
              $ref: '#/components/schemas/Meta'
    ReceiverPagedAuditLogsEnvelope:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelopeBase'
        - type: object
          required: [data, meta]
          properties:
            data:
              $ref: '#/components/schemas/ReceiverPagedAuditLogsResponse'
            meta:
              $ref: '#/components/schemas/Meta'
...
    SuccessEnvelopeBase:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Operation completed successfully.
...
    AuditLogDto:
      type: object
      required: [id, cardId, transactionAmount, decision, createdAt]
      properties:
        id:
          type: integer
          example: 90
        cardId:
          type: integer
          nullable: true
          example: 31
        transactionAmount:
          type: number
          format: decimal
          example: 25.5
        merchantName:
          type: string
          nullable: true
          example: Riyadh Taxi
        mcc:
          type: string
          nullable: true
          example: '4121'
        decision:
          $ref: '#/components/schemas/DecisionStatus'
        reason:
          type: string
          nullable: true
          example: "Transaction declined: insufficient remaining balance."
        createdAt:
          type: string
          format: date-time
          example: '2026-05-01T02:00:00Z'
        city:
          type: string
          nullable: true
          example: Riyadh
        country:
          type: string
          nullable: true
          description: Present only if the model exposes it.
          example: SA
        externalId:
          type: string
          nullable: true
          description: Present only if the model exposes it.
          example: ext_123
        entityId:
          type: integer
          nullable: true
          description: Logical entity id, commonly the intent id.
          example: 21
        action:
          type: string
          nullable: true
          example: authorization
        occurredAt:
          type: string
          format: date-time
          nullable: true
          example: '2026-05-01T02:00:00Z'
...
    ReceiverPagedAuditLogsResponse:
      type: object
      required: [receiverUserId, total, limit, offset, logs]
      properties:
        receiverUserId:
          type: integer
          example: 2
        total:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        offset:
          type: integer
          example: 0
        logs:
          type: array
          items:
            $ref: '#/components/schemas/AuditLogDto'
    UserTransactionsResponse:
      type: object
      required: [userId, total, limit, offset, logs]
      properties:
        userId:
          type: integer
          example: 1
        total:
          type: integer
          example: 1
        limit:
          type: integer
          example: 50
        offset:
          type: integer
          example: 0
        logs:
          type: array
          items:
            $ref: '#/components/schemas/AuditLogDto'
...
    UserLatestActivitiesResponse:
      type: object
      required: [userId, total, limit, offset, filters, summary, items]
      properties:
        userId:
          type: integer
          example: 1
        total:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        offset:
          type: integer
          example: 0
        filters:
          $ref: '#/components/schemas/UserLatestActivitiesAppliedFilters'
        summary:
          $ref: '#/components/schemas/UserLatestActivitiesSummary'
        items:
          type: array
          items:
            $ref: '#/components/schemas/UserActivityItem'
    UserLatestActivitiesAppliedFilters:
      type: object
      required: [role, includeInfo]
      properties:
        decision:
          $ref: '#/components/schemas/DecisionStatus'
        action:
          type: string
          nullable: true
          example: authorization
        entityType:
          type: string
          nullable: true
          description: Echo of query `entityType` when provided.
          enum: [profile, intent, virtual_card, transaction]
        outcome:
          type: string
          nullable: true
          description: Echo of query `status` (JSON field `outcome`).
          enum: [success, failed, info]
        cardId:
          type: integer
          nullable: true
          example: 31
        intentId:
          type: integer
          nullable: true
          example: 21
        fromUtc:
          type: string
          format: date-time
          nullable: true
        toUtc:
          type: string
          format: date-time
          nullable: true
        merchant:
          type: string
          nullable: true
          example: Taxi
        mcc:
          type: string
          nullable: true
          example: '4121'
        city:
          type: string
          nullable: true
          example: Riyadh
        minAmount:
          type: number
          format: decimal
          nullable: true
          example: 10.0
        maxAmount:
          type: number
          format: decimal
          nullable: true
          example: 100.0
        role:
          $ref: '#/components/schemas/ActivityRole'
        includeInfo:
          type: boolean
          example: true
    UserLatestActivitiesSummary:
      type: object
      required: [approvedCount, declinedCount, infoCount, approvedSpendTotal, declinedAmountTotal, distinctCards, distinctIntents]
      properties:
        approvedCount:
          type: integer
          example: 1
        declinedCount:
          type: integer
          example: 0
        infoCount:
          type: integer
          example: 0
        approvedSpendTotal:
          type: number
          format: decimal
          example: 25.5
        declinedAmountTotal:
          type: number
          format: decimal
          example: 0.0
        distinctCards:
          type: integer
          example: 1
        distinctIntents:
          type: integer
          example: 1
...